// src/openapi-client/models/DtsTemplateVO.ts
export interface DtsTemplateVO {
    /**
     * 
     * @type {string}
     * @memberof DtsTemplateVO
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsTemplateVO
     */
    description?: string;
    /**
     * 
     * @type {EntityState}
     * @memberof DtsTemplateVO
     */
    state?: EntityState;
    /**
     * 
     * @type {Array<EntityState>}
     * @memberof DtsTemplateVO
     */
    allowedTransitionStates?: Array<EntityState>;
    /**
     * 
     * @type {string}
     * @memberof DtsTemplateVO
     */
    label?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsTemplateVO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsTemplateVO
     */
    name?: string;
    /**
     * 
     * @type {DtsType}
     * @memberof DtsTemplateVO
     */
    type?: DtsType;
    /**
     * 
     * @type {string}
     * @memberof DtsTemplateVO
     */
    yaml?: string;
    /**
     * 
     * @type {Date}
     * @memberof DtsTemplateVO
     */
    createdTs?: Date;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof DtsTemplateVO
     */
    defaultDeploymentConfig?: { [key: string]: string; };
}